* Gebruik int voor width en height, niet ssize_t
git-svn-id: https://svn.il.fontys.nl/usersvn/fishtank/svn/trunk@318 bb1b31de-baa9-489e-af1c-03370f2903b8

+++ fishtank/main.cpp
@@ -572,7 +572,7 @@ static void DrawBackground(CAMERA camera)
 	glColor3f(1,1,1);
 }
 
-void render(AquariumController& aquariumController, CAMERA camera, int x, int y, ssize_t port_width, ssize_t port_height, const Eigen::Vector3d& area_size, const Eigen::Vector2d& facePosition = Eigen::Vector2d(0.5, 0.5))
+void render(AquariumController& aquariumController, CAMERA camera, int x, int y, int port_width, int port_height, const Eigen::Vector3d& area_size, const Eigen::Vector2d& facePosition = Eigen::Vector2d(0.5, 0.5))
 {
 	glViewport(x, y, port_width, port_height);
